data-pipeline/requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3